+++ Assets/DeviceID.txt
@@ -1,1 +1,1 @@
-﻿1
+﻿3
