+++ Assets/Scripts/SyncDataControl.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 3ad6984614164f14b9706c980b58063c
+timeCreated: 1506501502
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
